\name{r_pac_sum-package}
\alias{r_pac_sum-package}
\alias{r_pac_sum}
\docType{package}
\title{
\packageTitle{r_pac_sum}
}
\description{
\packageDescription{r_pac_sum}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{r_pac_sum}
\packageIndices{r_pac_sum}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{r_pac_sum}

Maintainer: \packageMaintainer{r_pac_sum}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
